+++ src/main/java/cn/vito/coding/check/po/TeacherAndAcademy.java
@@ -28,7 +28,7 @@ public class TeacherAndAcademy {
 	private String tenm;// 1000m
 	private Integer sit_ups;// 仰卧起坐
 	private Integer pull_up;// 引体向上
-	private Integer score;// 成绩
+	private Double score;// 成绩
 	private String first_check;// 第一次审核
 	private String second_check;// 第二次审核
 
@@ -38,7 +38,7 @@ public class TeacherAndAcademy {
 	public TeacherAndAcademy(String id, String name, String gender, String school, Integer grade, String classes,
 			Integer duration, String state, String teacher, Integer year, Integer height, Double weight,
 			Integer vital_capacity, Double fivem, Double long_jump, Double reach, String eightm, String tenm,
-			Integer sit_ups, Integer pull_up, Integer score, String first_check, String second_check) {
+			Integer sit_ups, Integer pull_up, Double score, String first_check, String second_check) {
 		this.id = id;
 		this.name = name;
 		this.gender = gender;
@@ -224,11 +224,11 @@ public class TeacherAndAcademy {
 		this.pull_up = pull_up;
 	}
 
-	public Integer getScore() {
+	public Double getScore() {
 		return score;
 	}
 
-	public void setScore(Integer score) {
+	public void setScore(Double score) {
 		this.score = score;
 	}
